requirements.txt
datasets==3.6.0
transformers==4.53.0
torch>=2.2.0
scikit-learn>=1.5.0
sklearn-crfsuite>=0.5.0
seqeval>=1.2.2
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
tqdm>=4.66.0
joblib>=1.4.0
pytorch-crf>=0.7.2
accelerate>=0.30.0
requests>=2.32.0